moved evaluator to internal folder so it can be used in multiple cmd/examples
