+++ cmd/server/main.go
@@ -3,29 +3,13 @@ package main
 import (
 	"encoding/json"
 	"net/http"
-	"strconv"
-	"strings"
 
-	"github.com/pBread/sms-content-moderator/internal/blacklist"
-	"github.com/pBread/sms-content-moderator/internal/llm"
+	"github.com/pBread/sms-content-moderator/internal/evaluator"
 	"github.com/pBread/sms-content-moderator/internal/logger"
 )
 
 type RequestBody struct {
 	Message string `json:"Message"`
-}
-
-type Evaluation struct {
-	Status    string `json:"status"`
-	Key       string `json:"key"`
-	Policy    string `json:"policy"`
-	Tier      int    `json:"tier"`
-	Reasoning string `json:"reasoning"`
-}
-
-type Response struct {
-	Status      string       `json:"status"`
-	Evaluations []Evaluation `json:"evaluations"`
 }
 
 func main() {
@@ -47,80 +31,10 @@ func unauthenticatedHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// checks message for blacklist entry matches
-	// returns []"{tier}-{policy}", e.g. ["0-profanity", "1-gambling"]
-	blacklistMatches := blacklist.CheckContent(reqBody.Message)
-
-	evaluations := []Evaluation{}
-	overallStatus := "pass"
-
-	tier0Present := false
-
-	for _, match := range blacklistMatches {
-		split := strings.Split(match, "-")
-		tier, _ := strconv.Atoi(split[0])
-		policy := split[1]
-
-		if tier == 0 {
-			evaluations = append(evaluations, Evaluation{
-				Status:    "is-violation",
-				Key:       match,
-				Policy:    policy,
-				Tier:      tier,
-				Reasoning: "Tier 0 blacklist entry was matched, which is automatically a policy violation.",
-			})
-			overallStatus = "fail"
-			tier0Present = true
-		}
-	}
-
-	if tier0Present {
-		// generate 'not-evaluated' records for Tier 1 if Tier 0 is present
-		for _, match := range blacklistMatches {
-			split := strings.Split(match, "-")
-			tier, _ := strconv.Atoi(split[0])
-			if tier == 1 {
-				evaluations = append(evaluations, Evaluation{
-					Status:    "not-evaluated",
-					Key:       match,
-					Policy:    split[1],
-					Tier:      tier,
-					Reasoning: "Message content included a Tier 0 blacklist violation and there is no reason to evaluate Tier 1 policies.",
-				})
-			}
-		}
-	} else if overallStatus == "pass" {
-		// evaluate Tier 1 if no Tier 0 is present
-		prompt, err := llm.BuildPrompt(reqBody.Message, blacklistMatches)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		// sends request to LLM for evaluation
-		llmViolations, err := llm.EvalPolicyViolation(prompt)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		for _, violation := range llmViolations {
-			evaluations = append(evaluations, Evaluation{
-				Status:    violation.Status,
-				Key:       "1-" + violation.Policy,
-				Policy:    violation.Policy,
-				Tier:      1,
-				Reasoning: violation.Reasoning,
-			})
-			if violation.Status == "is-violation" {
-				overallStatus = "fail"
-			}
-		}
-	}
-
-	response := Response{
-		Status:      overallStatus,
-		Evaluations: evaluations,
+	response, err := evaluator.EvaluateContent(reqBody.Message)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
 	}
 
 	w.Header().Set("Content-Type", "application/json")
